code refact clean up

+++ src/Moonglade.Auth/DeleteAccountCommand.cs
@@ -13,9 +13,6 @@ public class DeleteAccountCommandHandler : AsyncRequestHandler<DeleteAccountComm
     protected override async Task Handle(DeleteAccountCommand request, CancellationToken ct)
     {
         var account = await _repo.GetAsync(request.Id, ct);
-        if (account != null)
-        {
-            await _repo.DeleteAsync(request.Id, ct);
-        }
+        if (account != null) await _repo.DeleteAsync(request.Id, ct);
     }
 }
